+++ src/app.js
@@ -7,7 +7,7 @@ const filesRouter = require('./middlewares/routers/files_router.js');
 const serversRouter = require('./middlewares/routers/servers_router.js');
 const usersRouter = require('./middlewares/routers/users_router.js');
 const ConnectionPoolFactory = require('./lib/factories/connection_pool_factory.js');
-const errorMiddleware = require('./middlewares/error-middleware');
+const errorMiddleware = require('./middlewares/error_middleware.js');
 
 const logger = new Logger();
 const postgrePool = new ConnectionPoolFactory(logger).createPool();
